+++ src/System/IO/Streams/Core.hs
@@ -17,6 +17,7 @@ module System.IO.Streams.Core
  , unRead
  , peek
  , write
+ , writeTo
  , atEOF
 
    -- * Connecting streams together